refactor(search): split 588-line god component into focused modules
search/page.tsx: 588 lines → 52 lines (orchestrator only)

Extracted into:
- useSearch hook: all state, fetches, and save logic
- SearchHeader: form + quick filter pills
- DiscoverSection: genre carousels + social recommendations
- SearchResults: results grid, loading skeletons, empty state
- search-mappers: type definitions + DTO mappers + save helper

Each module has a single responsibility. The page is now
a thin composition layer that wires components to the hook.
